+++ src/ai_bom_generator/collectors/dependency_parsers/__init__.py
@@ -6,6 +6,7 @@ from ai_bom_generator.collectors.dependency_parsers.common import (
     ParserLimits,
 )
 from ai_bom_generator.collectors.dependency_parsers.conda_lock import parse_conda_lock
+from ai_bom_generator.collectors.dependency_parsers.poetry_lock import parse_poetry_lock
 from ai_bom_generator.collectors.dependency_parsers.requirements import parse_requirements
 from ai_bom_generator.collectors.dependency_parsers.uv_lock import parse_uv_lock
 
@@ -16,6 +17,7 @@ __all__ = [
     "DependencyParseResult",
     "ParserLimits",
     "parse_conda_lock",
+    "parse_poetry_lock",
     "parse_requirements",
     "parse_uv_lock",
 ]